lis2dw12: Making pdd->int_enable 16 bit mask
- Use pdd->int_enable as a 16 bit mask to track interrupts
  being enabled as opposed to a counter

+++ hw/drivers/sensors/lis2dw12/include/lis2dw12/lis2dw12.h
@@ -247,7 +247,7 @@ struct lis2dw12_pdd {
     /* Inetrrupt state */
     struct lis2dw12_int *interrupt;
     /* Interrupt enabled flag */
-    uint8_t int_enable;
+    uint16_t int_enable;
 };
 
 struct lis2dw12 {
